Keep orchestrator under 1k after priority-adopt seal paths (#222)
Move sealed inline/pool execute helpers onto the scope-lease mixin so the
DAG scheduler file stays under the 1000-line bar.
