+++ packages/api/api/compute/orchestrator.py
@@ -671,99 +671,6 @@ class ComputeOrchestrator(OrchestratorScopeLeaseMixin):
         for item in pending:
             self._run_inline_outside_lock(item)
 
-    def _run_inline_outside_lock(self, pending: _PendingInlineExecution) -> None:
-        node = pending.node
-        try:
-            builder = pending.registration.build_step_job_wire[pending.step.step_kind]
-            job_wire = builder(
-                node.scope,
-                dependency_outputs=pending.dependency_outputs,
-                ctx=self._cached_ctx,
-            )
-        except BaseException as exc:
-            with self._condition:
-                self._metrics.inline_executions += 1
-                self._observers.notify_step_complete(
-                    node,
-                    pending.step.step_kind,
-                    surface="inline",
-                    terminal_state="failed",
-                )
-                self._fail_node(node, exc)
-            return
-        # Seal after job-wire build so a higher-priority peer can adopt during
-        # that window.
-        if not self._seal_scope_lease_or_park(node, pending.step):
-            return
-        self._observers.notify_inline_start(node, pending.step.step_kind)
-        try:
-            result_wire = pending.registration.run_step[pending.step.step_kind](job_wire)
-        except BaseException as exc:
-            with self._condition:
-                self._metrics.inline_executions += 1
-                self._observers.notify_step_complete(
-                    node,
-                    pending.step.step_kind,
-                    surface="inline",
-                    terminal_state="failed",
-                )
-                self._fail_node(node, exc)
-            return
-        with self._condition:
-            self._metrics.inline_executions += 1
-            self._observers.notify_step_complete(
-                node,
-                pending.step.step_kind,
-                surface="inline",
-                terminal_state="success",
-            )
-            self._after_step_success(node, result_wire)
-
-    def _flush_pending_pool_submissions(
-        self,
-        pending: tuple[_PendingPoolSubmission, ...],
-    ) -> None:
-        """Build job wires and submit prepared pool work without the orchestrator lock."""
-        if not pending:
-            return
-        if self._pool_submitter is None:
-            raise RuntimeError("pool_submitter is not configured")
-        for submission in pending:
-            node = submission.node
-            step = submission.step
-            try:
-                if step.backend in {"interpreter", "process"}:
-                    builder = submission.registration.build_step_job_wire[step.step_kind]
-                    job_wire = builder(
-                        node.scope,
-                        dependency_outputs=submission.dependency_outputs,
-                        ctx=self._cached_ctx,
-                    )
-                    if not self._seal_scope_lease_or_park(node, step):
-                        continue
-                    run_step = submission.registration.run_step[step.step_kind]
-                    self._pool_submitter(
-                        node,
-                        step,
-                        job_wire=job_wire,
-                        run_step=run_step,
-                    )
-                else:
-                    if not self._seal_scope_lease_or_park(node, step):
-                        continue
-                    self._pool_submitter(node, step)
-                self._metrics.pool_submissions += 1
-            except BaseException as exc:
-                with self._condition:
-                    if node.state == "running":
-                        self._observers.notify_step_complete(
-                            node,
-                            step.step_kind,
-                            surface="pool",
-                            terminal_state="failed",
-                        )
-                        self._fail_node(node, exc)
-
     def _dequeue_dispatchable_ready_node(
         self,
     ) -> tuple[ComputeScope | None, ComputeNodeRun | None]:
